requirements.txt
beautifulsoup4==4.14.3
openai==2.9.0
pdfplumber==0.11.8
python-dotenv==1.2.1
Requests==2.32.5
schedule==1.2.2
streamlit==1.52.1
